pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biomedtools"
version = "0.1.0"
description = "轻量级生物医药计算工具库 - 零深度学习依赖"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "BioMedTools Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
]

[project.optional-dependencies]
rdkit = ["rdkit>=2022.03.1"]
biopython = ["biopython>=1.79"]
pymol = ["pymol-open-source"]
all = ["rdkit>=2022.03.1", "biopython>=1.79", "imageio>=2.9.0"]

[project.urls]
Homepage = "https://github.com/yourusername/biomedtools"
Repository = "https://github.com/yourusername/biomedtools"

[tool.setuptools.packages.find]
where = ["."]
include = ["biomedtools*"]
